Client - extract markdown type in its own package
