+++ client/src/views/Alias.vue
@@ -32,7 +32,8 @@ import {getDataItemByTypeHash} from 'dtype-core';
 import {TYPE_PREVIEW} from '../utils.js';
 
 import AliasSelector from '@/packages/alias/components/AliasSelector';
-import MarkdownRenderer from '@/packages/markdown/components/MarkdownRenderer';
+import MarkdownRenderer from 'dtype-markdown-ui';
+import 'dtype-markdown-ui/dist/dtype-markdown-ui.css';
 
 // http://192.168.1.140:8080/#/alias?alias=markdown.article1
 // http://192.168.1.140:8080/#/alias?alias=markdown.article1
