+++ packages/dtype-markdown-ui/index.js
@@ -1,0 +1,9 @@
+// import MarkdownRenderer from './src/components/MarkdownRenderer.vue';
+export * from './src/components/MarkdownRenderer.vue';
+
+// export default {
+//   install(Vue) {
+//     // Let's register our component globally
+//     Vue.component('markdown-renderer', MarkdownRenderer);
+//   },
+// };
