+++ packages/dtype-markdown-ui/vue.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  configureWebpack: {
+    output: {
+      libraryExport: 'default',
+    },
+    // resolve: {
+    //   symlinks: false,
+    // },
+  },
+};
