--- packages/dtypes-ui/dtype-markdown/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "dtype-markdown",
-  "version": "0.0.1",
-  "description": "dType markdown type",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/pipeos-one/dType.git"
-  },
-  "author": "",
-  "license": "GPL-3.0",
-  "bugs": {
-    "url": "https://github.com/pipeos-one/dType/issues"
-  },
-  "homepage": "https://github.com/pipeos-one/dType#readme"
-}
